Refactored code to fix the layout of everything.
