--- CS560Project/dependencies/map/bishopsmap.tmx
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<map version="1.0" orientation="orthogonal" renderorder="right-down" width="50" height="50" tilewidth="32" tileheight="32" nextobjectid="1">
- <tileset firstgid="1" name="tilesground" tilewidth="32" tileheight="32" tilecount="192" columns="16">
-  <image source="tilesimage.png" width="512" height="384"/>
- </tileset>
- <layer name="Background" width="50" height="50">
-  <data encoding="base64" compression="gzip">
-   H4sIAAAAAAAAC+3DAQkAAAwEofv+pZdjoOCqqaqqqqqqqqqq6vMHV4Re6RAnAAA=
-  </data>
- </layer>
- <layer name="Foreground" width="50" height="50">
-  <data encoding="base64" compression="gzip">
-   H4sIAAAAAAAAC+2XSw6AIBQDOSb3P4l74gLh9bUxM4k7oR0CfsYAAADQMt0FilB5dK8PHp55u/PwyMrDIysPj6w8PLLyHB6KTMf35188FOAB4CNx3550SvM4fX8lenSOU3HjkeJy2yXJ43Z8gktFh794VM7jzp/Dt8cUuQ4X5X9op4s6q8Ona71U58Z1Hqtync+TlTn2+3y518na8+0CANjhAazUa7UQJwAA
-  </data>
- </layer>
-</map>
