+++ README.md
@@ -1,0 +1,10 @@
+<h1> Treasure Trail </h1>
+
+<p> This is a game where the user runs around Bishop's
+university campus looking for clues that will lead the user
+to the hidden Bishop's treasure.</p>
+<br/>
+<p>We are using <a href="http://slick.ninjacave.com/">Slick2D</a> as a
+Java game library.</p>
+
+&copy; TreasureTrail 2017
